# internal/app/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A RESTful API for a pair-based online quiz game.\nPlayers compete head-to-head answering 5 questions. Super admin manages questions and users.",
        "title": "Online Quiz Game API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/api/pair-game-quiz/pairs/connection": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "If a pending game exists the calling user is added as second player and the game starts. Otherwise a new pending game is created.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz/Game"
                ],
                "summary": "Join or create a quiz game",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.GameResponse"
                        }
                    },
                    "403": {
                        "description": "Already in an active game",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/pair-game-quiz/pairs/my-current": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the game in status PendingSecondPlayer or Active. Returns 404 when no active game.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz/Game"
                ],
                "summary": "Get the calling user's current active game",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.GameResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/pair-game-quiz/pairs/my-current/answers": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submits an answer to the next unanswered question in the current active game. One attempt per question. Sequential order.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz/Game"
                ],
                "summary": "Submit an answer for the current question",
                "parameters": [
                    {
                        "description": "Answer",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.AnswerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.GameResponse"
                        }
                    },
                    "403": {
                        "description": "No active game or all questions answered",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/pair-game-quiz/pairs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the game in any status. The calling user must be a participant.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Quiz/Game"
                ],
                "summary": "Get a game by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Game ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.GameResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/testing/all-data": {
            "delete": {
                "description": "Truncates all application tables. For testing purposes only.",
                "tags": [
                    "Testing"
                ],
                "summary": "Delete all data",
                "responses": {
                    "204": {
                        "description": "All data deleted"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_pkg_apperrors.AppError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.AnswerRequest": {
            "type": "object",
            "required": [
                "answer"
            ],
            "properties": {
                "answer": {
                    "type": "string"
                }
            }
        },
        "github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.AnswerViewResponse": {
            "type": "object",
            "properties": {
                "addedAt": {
                    "type": "string"
                },
                "answerStatus": {
                    "type": "string"
                },
                "questionId": {
                    "type": "string"
                }
            }
        },
        "github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.GameResponse": {
            "type": "object",
            "properties": {
                "finishGameDate": {
                    "type": "string"
                },
                "firstPlayerProgress": {
                    "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.PlayerProgressResponse"
                },
                "id": {
                    "type": "string"
                },
                "pairCreatedDate": {
                    "type": "string"
                },
                "questions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.QuestionViewResponse"
                    }
                },
                "secondPlayerProgress": {
                    "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.PlayerProgressResponse"
                },
                "startGameDate": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.PlayerInfoResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                }
            }
        },
        "github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.PlayerProgressResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.AnswerViewResponse"
                    }
                },
                "player": {
                    "$ref": "#/definitions/github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.PlayerInfoResponse"
                },
                "score": {
                    "type": "integer"
                }
            }
        },
        "github_com_muzyk0_online-quiz-game_internal_domain_game_delivery_http_dto.QuestionViewResponse": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "github_com_muzyk0_online-quiz-game_internal_pkg_apperrors.AppError": {
            "type": "object",
            "properties": {
                "details": {
                    "description": "Details contains field-level validation errors (optional).",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "description": "Message is the safe message sent to the client.",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}